Add Elasticsearch tmpfs image
Our tests hit Elasticsearch quite hard. In fact, hard enough to use up
all burstable IOPS on the EC2 gp2 type EBS volume. To scale up disks,
you have to scale up the amount of storage you set them with - which is
all just more cost. This is a workaround for better performance (and so
we don't spend so much money).

To deliver better speed for MySQL we run it in tmpfs - this works really
well - but we were unable to do this for Elasticsearch because
Elasticsearch itself has to run as the "elasticsearch" user. When you
mount tmpfs, it mounts it as root, and the "elasticsearch" user does not
have permission to write to it.

This commit ammends the entry script slightly by setting the default
user as root; I feel this is OK since we're only using this in a testing
environment. It then chowns the volume after it's been mounted, and
starts the process as the "elasticsearch" user.