Update Drone file to build and push to Docker
At the moment, the futurelearn-drone container is a key container as we
are using it in Drone builds. It would be useful to automate the build
and push of this container for Dockerhub.

I have removed the other builds from the CI configuration as we are not
really using them yet. Further images should be added to both stages
(lint and build) as parallel steps.

I've also added the Slack notification step, and added the webhook as a
secret since this is a public repo.