Add zip
We were seeing ambigious errors about a file not being found by spec
tests. It started happening roughly after I updated the image.

There is a class which requires the zip package:

https://github.com/Futurelearn/futurelearn/blob/cc68e8da8f43850c5c3b6d655ddfc4548198f6a1/lib/future_learn/print_statements_batcher.rb#L8

I do recall seeing this issue before, but I did not change anything with
the image in recent changes.

One explanation could be adding changes and not pushing the changes to
Git.