Add parameter-store-exec to futurelearn-ruby
Using parameter-store-exec we can load in values from the parameter
store when starting up the container: https://github.com/cultureamp/parameter-store-exec

This is similar to [chamber](https://github.com/segmentio/chamber), but
unfortunately this does not allow path hierarchies which we require.

Include the binary with the base container to save on application build
time.

I am also removing pry because it was added for development reasons, but
we likely won't use it and it just increases image size.

This commit also updates Drone to push the latest image of both
futurelearn-drone and futurelearn-ruby to Dockerhub, but only on the
master branch. I do not think we need branch builds of these images (and
if we do, they can be done by hand).