Add warehouse-drone image
This is mostly based on the futurelearn-drone image with an adjusted set
of dependencies, and explicitly declared locale settings.

The locale needs to be set to something UTF-8 compatible as the test
suite exercises passing non-ASCII data to the shell. The ruby docker
hub page documents this setting under the heading "Encoding"[1]

[1]: https://hub.docker.com/_/ruby/