Add fix for postgresql-client
On the Ruby:2.5.3-slim image, the manpage directories are not present.
This causes the postgresql-client debian package to fail installation
because these directories are required. I am not sure why this has only
started to happen recently; either something has changed in the Ruby
slim image (or the underlying Debian slim image), or the
postgresql-client package.

Either way, the quickest fix is to create these directories. I did
consider using the non-slim Ruby image, but this is 8x bigger so opted
to create the directories, even if it's a little hacky.