Remove parameter-store-exec entirely
I've replaced parameter-store-exec with a dozen lines of Ruby in the
application itself.

I disliked having an external dependency to run the image, and it
doesn't do anything particularly special.

Having the load inside the boot of the app means we know exactly how the
environment is being set, and also gives us better coverage for testing
the process.