Create alpine version of futurelearn-ruby
It's actually easier to install the required dependencies because the
community have created a bunch of useful ones.

geoipupdate was the only stickler; it needs to be compiled with
CGO_ENABLED otherwise it doesn't use statically linked libraries which
means it doesn't work.

We can make use of multi stage builds to compile the latest version of
geoipupdate.

I've namespaced it so it's not built by default while we're testing.