Setup clamav
This adds a new docker image for running the clamav daemon, and installs
the clamav client in the `futurelearn-ruby` image.

The clamav client communicates with the daemon over TCP so we need to
configure the `TCPAddr` setting. This will be a different address when
testing locally versus ECS/EC2 mode and Fargate. As such, we can use the
`CLAM_DAEMON_ADDRESS` environment variable when running the container.

We're not running `freshclam` [1] as a daemon so we should rebuild the
`clamav` image on a regular basis. We might need to force rebuild it
without a cache, as docker might not detect a change in that command
even though there'll be new definitions.

[1]: https://www.clamav.net/documents/freshclam-faq