Create small backup tool
This is a small `mysqldump` backup tool which runs `mysqldump` against
an RDS instance, using IAM authentication[1], compressing the data using
[`pigz`](https://zlib.net/pigz/) and streaming the data into an S3
bucket using [`awscli`](https://aws.amazon.com/cli/).

I've written it to replace the use of running the MySQL backup within
our Rails app, which is not required since we do not do anything that
requires the Rails app to boot to backup the database.

It expects a `DATADOG_API_KEY` set since it will send an event to
Datadog; and I am forcing the use of the `ENVIRONMENT` environment
variable to ensure we have proper tagging on that event.

[1] https://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/UsingWithRDS.IAMDBAuth.html