Add clamdscan wrapper
We use this image as a sidecar next to an application container, and
mount the same volume into both containers. Occasionally we see errors
about "cannot access file" even when that file definitely exists on the
volume mount, and is normally shortly rectified immediately afterwards.

Our working theory is that there can sometimes be a small delay in
filesystem consistency between the two containers.

This adds a small wrapper that checks for the file existence for a
maximum of 10 seconds. This was the simplest way to resolve this rare
error.