Add healthcheck to clamav
This adds a built in healthcheck to the clamav container.

It takes at least 15 seconds to start up and load the databases, so
having a healthcheck to determine when it's healthy is the most useful.

We just want to know that the port is bound to the clam daemon, so
netcat is the easiest way to test this.