Do not install Node or Yarn
With Hypernova offloaded to a different container image which uses the
official node Docker image, we can remove Node and Yarn from this image.

This means we do not need to install some packages that were required
to install these deps.

We do need cURL in a later step, but we can install it all part of that
same step.