Add static Traefik reverse proxy
I'd like to test a small reverse proxy that proxies error pages to a
remote backend.

This is possible in NGINX, but it works out of the box for traefik, so
going to try with that.

I don't wish to use any dynamic configuration, since I'd like it to just
sit as a sidecar in front of the application, proxying all traffic to
the backend app, but intercepting errors, and passing them along to a
different location.

The default image doesn't come with envsubst which I'm using in the
entrypoint to generate the configuration, so we have to install it as
part of the image.

I am also explicitly disabling the api as part of the default command.