Explicitly set X-Forwarded-Proto
This header was being inserted by the proxy as "http". Rails uses this
header when the `force_ssl` option is set to `true`, and forces a
redirect.

We're terminating SSL at the load balancer, so this should be set as
`https`.

You cannot configure multiple types of middleware in a single map, so
I've had to specify two middlewares, one for errors and one for headers.

I've also renamed `catchall` to `wildcard` since I thought it was
clearer.