Add NGINX reverse proxy
I tried in vain to use Traefik as a small reverse proxy since it
forwarded error pages out of the box, and is built for cloud
environments.

Unfortunately I ran into issues with the `Authorization` header being
set when requesting the error page, and S3 failing since that header was
set.

This replaces the functionality with Good Ole NGINX, and strips the
`Authorization` header when we're passing on a request to an error page.

I've defined every single error page in a separate block with a separate
`upstream` block for the static pages, since I ran into issues trying to
be clever, where I had to set `resolver` configuration.

Basically, I got bored and decided not to be clever.