Set real IP configuration
We want to ensure the users remote IP is forwarded rather than an IP
from the load balancer or Cloudflare.

This users the set_real_ip_from config option[1] to ensure this happens.

We populate the configuration with Cloudflare addresses, and the private
network range used by our load balancers internally.

[1] https://nginx.org/en/docs/http/ngx_http_realip_module.html