Conditionally intercept errors
We need to conditionally intercept errors based upon the `Content-Type`
header. When our content is `text/html`, we would expect to return an
HTML error page.

Otherwise, we should not intercept errors, and instead pass back the
response by application.

I was hoping to use `proxy_intercept_errors` within an if statement, but
alas this not allowed in NGINX[1].

What we can do is use the `map` module[2] which is not dissimilar to
using an `if` statement.

We do a regex check based upon the built-in `$content_type` variable,
and if it matches `text/html`, then we set the `$intercept_errors`
variable to `intercept_errors_on`, otherwise we set it to the default of
`intercept_errors_off`.

We then use `rewrite`[3] to rewrite all requests to the location
matching the variable.

Using two location blocks, we can set `proxy_intercept_errors` to the
appropriate value.

We must use `$request_uri` as part of the `proxy_pass` argument
otherwise the `intercept_errors_on` or `intercept_errors_off` path will
also be included.

[1] https://nginx.org/en/docs/http/ngx_http_proxy_module.html#proxy_intercept_errors
[2] https://nginx.org/en/docs/http/ngx_http_map_module.html
[3] https://nginx.org/en/docs/http/ngx_http_rewrite_module.html