Load static pages on boot rather than proxy
We've run into a complicated issue where we need different responses
based upon different responses by the application.

I've tried several different ways to do this with NGINX, but it's
unfortunately proved too tricky.

At this point it feels like trying to smash a square into a round hole.

The complication is using `proxy_intercept_errors` to intercept all
errors, rather than letting the application display the errors. Our
current configuration only intercepts 3 different kinds of errors,
otherwise it is displayed by the application. This is possible since you
can declare to Rails where the static errors should exist.

My plan here is to load the errors onto the NGINX container, with
configuration to intercept a subset.

We will then mount the volume with the Rails container, so it should
have access to these error pages.